requirements.txt
flask
python-chess
gunicorn
